Move `FilteringGateway` from `server.entity` to `system.server`.
